requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
tensorflow>=2.10.0